Split connection.rs into files
Divide connection.rs into separate files for the command and reply
enums and their related code. Unfortunately, there's enough boilerplate
surrounding the reply that Git seems convinced that I renamed
connection.rs to reply.rs not mod.rs, but I suppose there's an argument
to be made either way.
