+++ src/connection/reply.rs
@@ -1,186 +1,4 @@
-use std::io;
-use std::io::prelude::*;
-use std::net;
-use std::str::FromStr;
-
-pub struct Connection {
-    reader: io::BufReader<net::TcpStream>,
-    writer: net::TcpStream,
-}
-
-impl<'a> Connection {
-    pub fn new(stream: net::TcpStream) -> Connection {
-        stream.set_nonblocking(true).unwrap();
-
-        let reader = io::BufReader::new(stream.try_clone().unwrap());
-
-        Connection {
-            reader,
-            writer: stream,
-        }
-    }
-
-    pub fn poll(&mut self) -> Option<Message> {
-        let mut buffer = String::new();
-
-        match self.reader.read_line(&mut buffer) {
-            Ok(len) => {
-                if len == 0 {
-                    panic!("Stream disconnected");
-                } else {
-                    // Truncate at the first control character (ie. CR/LF)
-                    buffer.truncate(buffer.find(char::is_control).unwrap_or(buffer.len()));
-                    split_server_name(&mut buffer);
-
-                    if let Some(command) = raw_to_command(&buffer) {
-                        println!("\x1B[94m<C {:?}\x1B[0m", command);
-                        Some(Message::Command(command))
-                    } else if let Some((reply_type, reply_body)) = raw_to_reply(&buffer) {
-                        println!("\x1B[92m<R {:?} {:?}\x1B[0m", reply_type, reply_body);
-                        Some(Message::Reply(reply_type, reply_body))
-                    } else {
-                        println!("\x1B[91m<? {}\x1B[0m", buffer);
-                        None
-                    }
-                }
-            }
-            Err(ref e) if e.kind() == io::ErrorKind::WouldBlock => None,
-            Err(e) => panic!("IO error: {}", e),
-        }
-    }
-
-    pub fn send_command(&mut self, command: Command) -> std::io::Result<()> {
-        let raw_command = command_to_raw(command);
-        self.send_command_raw(raw_command)
-    }
-
-    pub fn send_command_raw(&mut self, mut raw_command: String) -> std::io::Result<()> {
-        raw_command.push_str("\r\n");
-        print!(">> {}", raw_command);
-        self.writer.write(raw_command.as_bytes())?;
-        Ok(())
-    }
-}
-
-pub enum Message {
-    Command(Command),
-    Reply(ReplyType, String),
-}
-
-fn split_server_name(raw_message: &mut String) -> Option<String> {
-    if raw_message.starts_with(':') {
-        let slice_index = raw_message.find(' ')?;
-        let server_name = raw_message[1..slice_index].to_string();
-        raw_message.replace_range(..slice_index + 1, "");
-        Some(server_name)
-    } else {
-        None
-    }
-}
-
-fn raw_to_command(raw_command: &str) -> Option<Command> {
-    let command_parts: Vec<&str> = raw_command.split(' ').collect();
-
-    match command_parts.first()?.as_ref() {
-        "PASS" => {
-            if command_parts.len() == 2 {
-                Some(Command::Pass {
-                    password: command_parts[1].to_string(),
-                })
-            } else {
-                None
-            }
-        }
-        "NICK" => {
-            if command_parts.len() >= 2 && command_parts.len() <= 3 {
-                Some(Command::Nick {
-                    nickname: command_parts[1].to_string(),
-                    hopcount: match command_parts.get(2) {
-                        Some(n) => u8::from_str(n).ok(),
-                        None => None,
-                    },
-                })
-            } else {
-                None
-            }
-        }
-        "USER" => {
-            if command_parts.len() >= 5 {
-                Some(Command::User {
-                    username: command_parts[1].to_string(),
-                    mode: u8::from_str(command_parts[2]).unwrap_or(0),
-                    // part 3 is unused
-                    realname: command_parts[4..].join(" ").strip_prefix(":")?.to_string(),
-                })
-            } else {
-                None
-            }
-        }
-        "PING" => match command_parts.len() {
-            1 => Some(Command::Ping {
-                to: None,
-                from: None,
-            }),
-            2 => {
-                if Some(':') == command_parts[1].chars().nth(0) {
-                    Some(Command::Ping {
-                        to: None,
-                        from: Some(command_parts[1].get(1..).unwrap_or("").to_string()),
-                    })
-                } else {
-                    Some(Command::Ping {
-                        to: Some(command_parts[1].to_string()),
-                        from: None,
-                    })
-                }
-            }
-            3 => Some(Command::Ping {
-                to: Some(command_parts[2].to_string()),
-                from: Some(command_parts[1].to_string()),
-            }),
-            _ => None,
-        },
-        "PONG" => match command_parts.len() {
-            2 => Some(Command::Pong {
-                to: None,
-                from: command_parts[1].to_string(),
-            }),
-            3 => Some(Command::Pong {
-                to: Some(command_parts[2].to_string()),
-                from: command_parts[1].to_string(),
-            }),
-            _ => None,
-        },
-        _ => None,
-    }
-}
-
-fn command_to_raw(command: Command) -> String {
-    match command {
-        Command::Pass { password } => format!("PASS {}", password),
-        Command::Nick { nickname, hopcount } => match hopcount {
-            Some(hopcount) => format!("NICK {} {}", nickname, hopcount),
-            None => format!("NICK {}", nickname),
-        },
-        Command::User {
-            username,
-            mode,
-            realname,
-        } => format!("USER {} {} * :{}", username, mode, realname),
-        Command::Ping { to, from } => match (to, from) {
-            (Some(to), Some(from)) => format!("PING {} {}", from, to),
-            (Some(to), None) => format!("PING {}", to),
-            (None, Some(from)) => format!("PING :{}", from),
-            (None, None) => "PING".to_string(),
-        },
-        Command::Pong { to, from } => match to {
-            Some(to) => format!("PONG {} {}", from, to),
-            None => format!("PONG {}", from),
-        },
-    }
-}
-
-fn raw_to_reply(raw_reply: &str) -> Option<(ReplyType, String)> {
+pub fn from_raw(raw_reply: &str) -> Option<(ReplyType, String)> {
     let reply_parts: Vec<&str> = raw_reply.split(' ').collect();
 
     let (reply_prefix, reply_body_parts) = reply_parts.split_first()?;
@@ -331,43 +149,6 @@ fn raw_to_reply(raw_reply: &str) -> Option<(ReplyType, String)> {
     ))
 }
 
-#[derive(Debug)]
-pub enum Command {
-    // Connection registration
-    Pass {
-        password: String,
-    },
-    Nick {
-        nickname: String,
-        hopcount: Option<u8>,
-    },
-    User {
-        username: String,
-        mode: u8,
-        realname: String,
-    },
-    //Oper { user: String, password: String },
-    //Quit { message: Option<String> },
-
-    // Channel operations
-    //Join { channels: Vec<String>, keys: Vec<String> },
-    //Part { channels: Vec<String> },
-
-    // Sending messages
-    //Privmsg { receivers: Vec<Messageable>, message: String },
-    //Notice { receivers: Vec<Messageable>, message: String },
-
-    // Miscellaneous messages
-    Ping {
-        from: Option<String>,
-        to: Option<String>,
-    },
-    Pong {
-        from: String,
-        to: Option<String>,
-    },
-}
-
 #[derive(Hash, Eq, PartialEq, Debug)]
 pub enum ReplyType {
     RplWelcome,           // 001 - "Welcome to the Internet Relay Network
@@ -514,258 +295,17 @@ mod tests {
     use super::*;
 
     #[test]
-    fn command_to_raw_pass() {
-        assert_eq!(
-            "PASS mysecretpass",
-            command_to_raw(Command::Pass {
-                password: "mysecretpass".to_string(),
-            }),
-        );
-    }
-
-    #[test]
-    fn command_to_raw_nick() {
-        assert_eq!(
-            "NICK potato",
-            command_to_raw(Command::Nick {
-                nickname: "potato".to_string(),
-                hopcount: None,
-            }),
-        );
-        assert_eq!(
-            "NICK carrot 5",
-            command_to_raw(Command::Nick {
-                nickname: "carrot".to_string(),
-                hopcount: Some(5),
-            }),
-        );
-    }
-
-    #[test]
-    fn command_to_raw_user() {
-        assert_eq!(
-            "USER pjohnson 0 * :Potato Johnson",
-            command_to_raw(Command::User {
-                username: "pjohnson".to_string(),
-                mode: 0,
-                realname: "Potato Johnson".to_string(),
-            }),
-        );
-    }
-
-    #[test]
-    fn command_to_raw_ping() {
-        assert_eq!(
-            "PING",
-            command_to_raw(Command::Ping {
-                to: None,
-                from: None
-            }),
-        );
-        assert_eq!(
-            "PING :me",
-            command_to_raw(Command::Ping {
-                to: None,
-                from: Some("me".to_string()),
-            }),
-        );
-        assert_eq!(
-            "PING myserver",
-            command_to_raw(Command::Ping {
-                to: Some("myserver".to_string()),
-                from: None
-            }),
-        );
-        assert_eq!(
-            "PING me myserver",
-            command_to_raw(Command::Ping {
-                to: Some("myserver".to_string()),
-                from: Some("me".to_string()),
-            }),
-        );
-    }
-
-    #[test]
-    fn command_to_raw_pong() {
-        assert_eq!(
-            "PONG me",
-            command_to_raw(Command::Pong {
-                from: "me".to_string(),
-                to: None,
-            }),
-        );
-        assert_eq!(
-            "PONG me myserver",
-            command_to_raw(Command::Pong {
-                from: "me".to_string(),
-                to: Some("myserver".to_string()),
-            }),
-        );
-    }
-
-    #[test]
-    fn raw_to_command_pass() {
-        let command = raw_to_command("PASS mysecretpass");
-        if let Some(Command::Pass { password }) = command {
-            assert_eq!("mysecretpass", password);
-        } else {
-            panic!("Wrong type: {:?}", command);
-        }
-    }
-
-    #[test]
-    fn raw_to_command_nick() {
-        let command = raw_to_command("NICK somebody");
-        if let Some(Command::Nick { nickname, hopcount }) = command {
-            assert_eq!("somebody", nickname);
-            assert_eq!(None, hopcount);
-        } else {
-            panic!("Wrong type: {:?}", command);
-        }
-
-        let command = raw_to_command("NICK anybody 5");
-        if let Some(Command::Nick { nickname, hopcount }) = command {
-            assert_eq!("anybody", nickname);
-            assert_eq!(Some(5), hopcount);
-        } else {
-            panic!("Wrong type: {:?}", command);
-        }
-
-        let command = raw_to_command("NICK anybody potato");
-        if let Some(Command::Nick { nickname, hopcount }) = command {
-            assert_eq!("anybody", nickname);
-            assert_eq!(None, hopcount);
-        } else {
-            panic!("Wrong type: {:?}", command);
-        }
-    }
-
-    #[test]
-    fn raw_to_command_user() {
-        assert!(raw_to_command("USER pjohnson 0 *").is_none());
-        assert!(raw_to_command("USER pjohnson 0 * realname").is_none());
-
-        let command = raw_to_command("USER pjohnson 0 * :Potato Johnson");
-        if let Some(Command::User {
-            username,
-            mode,
-            realname,
-        }) = command
+    fn from_raw_welcome() {
+        if let Some((reply_type, reply_body)) =
+            from_raw("001 :Welcome to the Internet Relay Network nick!~username@host")
         {
-            assert_eq!("pjohnson", username);
-            assert_eq!(0, mode);
-            assert_eq!("Potato Johnson", realname);
+            assert_eq!(ReplyType::RplWelcome, reply_type);
+            assert_eq!(
+                ":Welcome to the Internet Relay Network nick!~username@host",
+                reply_body
+            );
         } else {
-            panic!("Wrong type: {:?}", command);
+            panic!("Wrong type: None");
         }
-    }
-
-    #[test]
-    fn raw_to_command_ping() {
-        let command = raw_to_command("PING");
-        if let Some(Command::Ping { to, from }) = command {
-            assert!(to.is_none());
-            assert!(from.is_none());
-        } else {
-            panic!("Wrong type: {:?}", command);
-        }
-
-        let command = raw_to_command("PING myserver");
-        if let Some(Command::Ping { to, from }) = command {
-            assert_eq!(Some("myserver".to_string()), to);
-            assert!(from.is_none());
-        } else {
-            panic!("Wrong type: {:?}", command);
-        }
-
-        let command = raw_to_command("PING me myserver");
-        if let Some(Command::Ping { to, from }) = command {
-            assert_eq!(Some("myserver".to_string()), to);
-            assert_eq!(Some("me".to_string()), from);
-        } else {
-            panic!("Wrong type: {:?}", command);
-        }
-
-        let command = raw_to_command("PING :me");
-        if let Some(Command::Ping { to, from }) = command {
-            assert!(to.is_none());
-            assert_eq!(Some("me".to_string()), from);
-        } else {
-            panic!("Wrong type: {:?}", command);
-        }
-
-        assert!(raw_to_command("PING a b c").is_none());
-    }
-
-    #[test]
-    fn raw_to_command_pong() {
-        let command = raw_to_command("PONG me");
-        if let Some(Command::Pong { to, from }) = command {
-            assert_eq!("me".to_string(), from);
-            assert!(to.is_none());
-        } else {
-            panic!("Wrong type: {:?}", command);
-        }
-
-        let command = raw_to_command("PONG me myserver");
-        if let Some(Command::Pong { to, from }) = command {
-            assert_eq!("me".to_string(), from);
-            assert_eq!(Some("myserver".to_string()), to);
-        } else {
-            panic!("Wrong type: {:?}", command);
-        }
-
-        assert!(raw_to_command("PONG").is_none());
-        assert!(raw_to_command("PONG a b c").is_none());
-    }
-
-    #[test]
-    fn split_server_name_with_name() {
-        let mut command =
-            ":irc.example.net 001 foo :Welcome to the Internet Relay Network".to_string();
-        let result = split_server_name(&mut command);
-
-        assert_eq!("001 foo :Welcome to the Internet Relay Network", command);
-        assert_eq!(Some("irc.example.net".to_string()), result);
-    }
-
-    #[test]
-    fn split_server_name_no_name() {
-        let mut command = "001 foo :Welcome to the Internet Relay Network".to_string();
-        let result = split_server_name(&mut command);
-
-        assert_eq!("001 foo :Welcome to the Internet Relay Network", command);
-        assert_eq!(None, result);
-    }
-
-    #[test]
-    fn split_server_name_missing_colon() {
-        let mut command =
-            "irc.example.net 001 foo :Welcome to the Internet Relay Network".to_string();
-        let result = split_server_name(&mut command);
-
-        assert_eq!(
-            "irc.example.net 001 foo :Welcome to the Internet Relay Network",
-            command
-        );
-        assert_eq!(None, result);
-    }
-
-    #[test]
-    fn split_server_name_server_only() {
-        let mut command = ":irc.example.net".to_string();
-        let result = split_server_name(&mut command);
-
-        assert_eq!(":irc.example.net", command);
-        assert_eq!(None, result);
-    }
-
-    #[test]
-    fn split_server_name_trailing_space() {
-        let mut command = ":irc.example.net ".to_string();
-        let result = split_server_name(&mut command);
-
-        assert_eq!("", command);
-        assert_eq!(Some("irc.example.net".to_string()), result);
     }
 }
